update: 将pickers换成无状态组件. 修复object render 报错.
